Cleaned up extra code.
Added more fine grain control over determining if this is the first time the user is starting the application. If it isn't and it is a new version, it will pull the old settings. Note, this still doesn't work if the user decides to change the location of the program. But it does at least go half way.