Added the ability to keep the shutdown notifications. Ultimately the timer is run through the task scheduler as a task that sets the timer onto the command and executes the task immediately. This will result in the timer notifying the user of a shutdown.
Note: This will result in possibly inaccurate "Time Remaining" measurements since there isn't a reliable method of checking for the time remaining.

Cleaned up the options class.

Fixed a bug where the window would not longer be the top window after saving or completing an action.

Upgraded the version to 2.1.0.0.